Extract useResourceBuyingStateMachine hook
